Change double to float for position variables

+++ Chemotaxis.pde
@@ -25,10 +25,10 @@ void setup () {
 int mouseMode = 0;
 
 class RPS {
-  double myX, myY;
+  float myX, myY;
   int mode = 1;
-  double left, right, upper, lower;
-  double size = 25;
+  float left, right, upper, lower;
+  float size = 25;
   
   RPS(int m) {
     myX = Math.random()*400;
@@ -45,11 +45,11 @@ class RPS {
   }
   
   void show() {
-    double mX = mouseX;
-    double mY = mouseY;
-    double a = mX - myX;
-    double b = mY - myY;
-    double c = Math.sqrt(Math.pow(a, 2)+Math.pow(b, 2));
+    float mX = mouseX;
+    float mY = mouseY;
+    float a = mX - myX;
+    float b = mY - myY;
+    float c = Math.sqrt(Math.pow(a, 2)+Math.pow(b, 2));
     if (mode == mouseMode && mousePressed) {
       myX += -a/a*2*(a/Math.pow(c, 2/3) + (a/Math.abs(a))*Math.random()*Math.pow(c, 1/2))/(30);
       myY += -b/b*2*(b/Math.pow(c, 2/3) + (b/Math.abs(b))*Math.random()*Math.pow(c, 1/2))/(30);
